Add tracing
- Since this is eventually going to be used by iced GUI, tracing seems a
  better choice than log

+++ src/main.rs
@@ -1,5 +1,6 @@
 use clap::Parser;
 use std::path::PathBuf;
+use tracing::{error, info};
 
 #[derive(Parser)]
 struct Opt {
@@ -9,5 +10,19 @@ struct Opt {
 
 fn main() {
     let opt = Opt::parse();
+
+    tracing_subscriber::fmt().init();
+
+    if !opt.img.exists() {
+        error!("Provided Image does not exist {:?}", opt.img);
+        return;
+    }
+
+    if !opt.dst.exists() {
+        error!("Provided destination does not exist {:?}", opt.dst);
+        return;
+    }
+
+    info!("Writing image {:?} to {:?}", opt.img, opt.dst);
     std::fs::copy(opt.img, opt.dst).unwrap();
 }
